requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
anthropic==0.25.6
mysql-connector-python==8.3.0
phpserialize==1.3
